+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-2.0.0 / WIP
+2.0.0 / 2020-10-15
 ------------------
 
 - Drop deprecated `.to_blob()` & `to_canvas()` (use `.toBlob()` & `toCanvas()`